Implement negate, multiply, divide, and magnitude on Vector3

+++ lib/src/main/java/com/gregjandl/raytracer/Utils.java
@@ -3,7 +3,7 @@ package com.gregjandl.raytracer;
 public class Utils {
   public static final float EPSILON = 0.0001f;
 
-  public static boolean aboutEqual(float lhs, float rhs) {
+  public static boolean aboutEqual(double lhs, double rhs) {
     return Math.abs(lhs - rhs) < EPSILON;
   }
